Déplace la méthode setEntreprise dans un hook
L'idée est de regrouper la logique métier dans un fichier unique séparé
des vues. Il reste un peu de travail pour généraliser ce principe.
